Remove arrow2 bindings
