+++ Cargo.toml
@@ -16,9 +16,7 @@ crate-type = ["cdylib", "rlib"]
 
 [features]
 default = ["all"]
-all = ["arrow1", "data", "data_type", "read_arrow_js", "table", "vector"]
-
-arrow1 = ["dep:arrow", "dep:arrow-schema"]
+all = ["data", "data_type", "read_arrow_js", "table", "vector"]
 
 # Include Data classes for contiguous Arrow memory
 data = []
@@ -35,7 +33,6 @@ table = []
 # Include Vector classes for chunked Arrow memory
 vector = []
 
-arrow2 = ["dep:arrow2"]
 debug = ["console_error_panic_hook"]
 
 [dependencies]
@@ -57,12 +54,8 @@ js-sys = "0.3.60"
 getrandom = { version = "0.2.6", features = ["js"] }
 thiserror = "1.0"
 
-arrow2 = { version = "0.18", optional = true, features = ["io_ipc"] }
-
-# Temporary for compatibility with geoarrow-rs until next release
-#
-arrow-schema = { version = "50", optional = true }
-arrow = { version = "50", default-features = false, optional = true, features = [
+arrow-schema = "50"
+arrow = { version = "50", default-features = false, features = [
   "ffi",
   "ipc",
 ] }
@@ -71,20 +64,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde-wasm-bindgen = "0.6"
 
 wasm-bindgen-futures = { version = "0.4.30", optional = true }
-
-[dependencies.web-sys]
-version = "0.3.4"
-features = [
-  'console',
-  'Headers',
-  'Request',
-  'RequestInit',
-  'RequestMode',
-  'Response',
-  'Window',
-  "Document",
-  "Element",
-]
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.31"
