+++ package.json
@@ -14,7 +14,7 @@
     "yarn": "3.6.3"
   },
   "devDependencies": {
-    "apache-arrow": "^14.0.1",
+    "apache-arrow": "^15.0.0",
     "gh-pages": "^6.0.0",
     "typedoc": "^0.25.1",
     "typescript": "^5.2.2"
