--- src/arrow2/mod.rs
@@ -1,12 +1,0 @@
-mod datatype;
-pub mod error;
-mod record_batch;
-mod schema;
-mod table;
-mod vector;
-
-pub use datatype::DataType;
-pub use record_batch::{FFIRecordBatch, RecordBatch};
-pub use schema::Schema;
-pub use table::{FFITable, Table};
-pub use vector::{FFIVector, Vector};
