--- src/arrow2/vector/mod.rs
@@ -1,5 +1,0 @@
-mod ffi;
-mod untyped;
-
-pub use ffi::FFIVector;
-pub use untyped::Vector;
