+++ src/arrow_js/data.rs
@@ -3,7 +3,7 @@ use arrow::buffer::Buffer;
 use arrow_schema::DataType;
 use wasm_bindgen::prelude::*;
 
-use crate::arrow1::arrow_js::r#type::{import_data_type, JSDataType};
+use crate::arrow_js::r#type::{import_data_type, JSDataType};
 
 #[wasm_bindgen]
 extern "C" {
