+++ src/arrow_js/type.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use arrow_schema::{DataType, UnionFields};
 use wasm_bindgen::prelude::*;
 
-use crate::arrow1::arrow_js::field::{import_field, JSField};
+use crate::arrow_js::field::{import_field, JSField};
 
 #[wasm_bindgen]
 extern "C" {
