+++ src/lib.rs
@@ -1,10 +1,26 @@
 use wasm_bindgen::prelude::*;
 
-#[cfg(feature = "arrow1")]
-pub mod arrow1;
+#[cfg(feature = "read_arrow_js")]
+pub mod arrow_js;
+#[cfg(feature = "data")]
+pub mod data;
+#[cfg(feature = "data_type")]
+pub mod datatype;
+pub mod error;
+pub mod ffi;
+pub mod field;
+pub mod record_batch;
+pub mod schema;
+#[cfg(feature = "table")]
+pub mod table;
+#[cfg(feature = "vector")]
+pub mod vector;
 
-#[cfg(feature = "arrow2")]
-pub mod arrow2;
+pub use error::ArrowWasmError;
+pub use field::Field;
+pub use record_batch::{FFIRecordBatch, RecordBatch};
+pub use schema::Schema;
+pub use table::{FFITable, Table};
 
 mod utils;
