+++ src/vector/ffi.rs
